+++ rand_distr/src/dirichlet.rs
@@ -92,7 +92,7 @@ where
 }
 
 #[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 struct DirichletFromBeta<F, const N: usize>
 where
     F: Float,