+++ src/distr/uniform_int.rs
@@ -20,7 +20,7 @@ use core::simd::prelude::*;
 #[cfg(feature = "simd_support")]
 use core::simd::{LaneCount, SupportedLaneCount};
 
-#[cfg(feature = "serde1")]
+#[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
 /// The back-end implementing [`UniformSampler`] for integer types.
@@ -60,7 +60,7 @@ use serde::{Deserialize, Serialize};
 ///
 /// [`Uniform`]: super::Uniform
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serde1", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct UniformInt<X> {
     pub(super) low: X,
     pub(super) range: X,