+++ rand_chacha/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- The `serde1` feature has been renamed `serde` (#1477)
+
 ## [0.9.0-alpha.1] - 2024-03-18
 
 ## [0.9.0-alpha.0] - 2024-02-18